juste comme ça. update

+++ routes/api.php
@@ -1,17 +1,14 @@
 <?php
 
-use App\Http\Controllers\Api\AuthController;
-use App\Http\Controllers\Api\CartController;
-use App\Http\Controllers\Api\CategoryController;
-use App\Http\Controllers\Api\OrderController;
-use App\Http\Controllers\Api\ProductController;
-use App\Http\Controllers\Api\UserController;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Route;
+    use App\Http\Controllers\Api\CartController;
+    use App\Http\Controllers\Api\CategorieController;
+    use App\Http\Controllers\Api\OrderController;
+    use App\Http\Controllers\Api\ProductController;
+    use App\Http\Controllers\Api\UserController;
+    use App\Http\Controllers\AuthController;
+    use Illuminate\Support\Facades\Route;
 
-Route::get('/', function () {
-    return view('welcome');
-});
+
 
 Route::prefix('carts')->group(function () {
     Route::get('/', [CartController::class, 'index']);
@@ -23,13 +20,13 @@ Route::prefix('carts')->group(function () {
 
 // Routes pour les utilisateurs
 Route::prefix('users')->group(function () {
-    Route::get('/', [UserController::class, 'index'])->name('users.index');
-    Route::get('/create', [UserController::class, 'create'])->name('users.create');
-    Route::get('/{user}/edit', [UserController::class, 'edit'])->name('users.edit');
-    Route::put('/{user}', [UserController::class, 'update'])->name('users.update');
-    Route::post('/', [UserController::class, 'store'])->name('users.store');
-    Route::get('/{user}', [UserController::class, 'show'])->name('users.show');
-    Route::delete('/{user}', [UserController::class, 'destroy'])->name('users.destroy');
+    Route::get('/', [UserController::class, 'index'])->name('users.index')->Middleware('auth:sanctum');
+    Route::get('/create', [UserController::class, 'create'])->name('users.create')->Middleware('auth:sanctum');
+    Route::get('/{user}/edit', [UserController::class, 'edit'])->name('users.edit')->Middleware('auth:sanctum');
+    Route::put('/{user}', [UserController::class, 'update'])->name('users.update')->Middleware('auth:sanctum');
+    Route::post('/', [UserController::class, 'store'])->name('users.store')->Middleware('auth:sanctum');
+    Route::get('/{user}', [UserController::class, 'show'])->name('users.show')->Middleware('auth:sanctum');
+    Route::delete('/{user}', [UserController::class, 'destroy'])->name('users.destroy')->Middleware('auth:sanctum');
 });
 //->Middleware('auth:sanctum')
 
@@ -44,11 +41,11 @@ route::prefix('/products')->group(function () {
 
 // Routes pour les catégories
 Route::prefix('/categorie')->group(function () {
-    Route::get('/', [CategoryController::class, 'index']);
-    Route::get('/{id}', [CategoryController::class, 'showProductList']);
-    Route::post('/', [CategoryController::class, 'create']);
-    Route::put('/{id}', [CategoryController::class, 'update']);
-    Route::delete('/{id}', [CategoryController::class, 'delete']);
+    Route::get('/', [CategorieController::class, 'index']);
+    Route::get('/{id}', [CategorieController::class, 'showCategorie']);
+    Route::post('/', [CategorieController::class, 'create']);
+    Route::put('/{id}', [CategorieController::class, 'update']);
+    Route::delete('/{id}', [CategorieController::class, 'delete']);
 });
 
 // Routes pour les paniers
@@ -63,10 +60,5 @@ Route::prefix('/orders')->group(function () {
     Route::delete('/{id}', [OrderController::class, 'destroy']);
 });
 
-Route::get('/login', [AuthController::class, 'login'])->name('auth.login');
-Route::post('/login', [AuthController::class, 'doLogin']);
-
-// Exemple de route pour obtenir l'utilisateur authentifié
-Route::get('/user', function (Request $request) {
-    return $request->user();
-})->Middleware('auth:sanctum');
+    Route::get('/login', [AuthController::class, 'login'])->name('auth.login');
+    Route::post('/login', [AuthController::class, 'doLogin']);
